Add default icons to the getPreview method
Feature:
Major update to the getPreview method. Now getPreview returns a preview element (if the file type allows it)
but also a file icon based on the given file extension. This icon is created as a canvas element and returned as one.
It's colors can be changed by passing color options to the constructor (options.iconColors).
getPreview's logic was also improved and a switch was added to more concisely handle the various methods for creating blobs from the given filetype.
Several helpers methods were added to draw the file icon in the canvas.

Documentaiton:
Updated readme

Chore:
-Added yalc.lock to npmignore
-removed unused nodule modules from package.json

Developemnt:
added 'build-dev' script, that builds unminifed fileselect script
and publishes node module to yalc